feat(pkg/server): support database loading/unloading

chore(pkg/server): de-duplicate db logging


feat(pkg/api): endpoint for listing databases with settings


chore(pkg/server): db listing with settings


test(pkg/server): unit testing db loading and closing


chore(pkg/server): set smaller systemid offset, used in deprecated endpoints


chore(pkg/client): check connection status when listing dbs


test(pkg/server): adjust unit testing after rebasing from master


test(pkg/server): unit testing dummy closed db


chore(pkg/api): auto-open, open and close db mgmt


chore(pkg/server): start/stop replication only when db is open


feat(pkg/client): db open and close


chore(pkg/api): direct use of autoOpen property

+++ pkg/server/user.go
@@ -150,7 +150,7 @@ func (s *ImmuServer) ListUsers(ctx context.Context, req *empty.Empty) (*schema.U
 	s.Logger.Debugf("ListUsers")
 
 	loggedInuser := &auth.User{}
-	var dbInd = int64(0)
+	var dbInd = 0
 	var err error
 	userlist := &schema.UserList{}
 
@@ -600,7 +600,7 @@ func (s *ImmuServer) addUserToLoginList(u *auth.User) {
 	s.userdata.Userdata[u.Username] = u
 }
 
-func (s *ImmuServer) getLoggedInUserdataFromCtx(ctx context.Context) (int64, *auth.User, error) {
+func (s *ImmuServer) getLoggedInUserdataFromCtx(ctx context.Context) (int, *auth.User, error) {
 	if sessionID, err := sessions.GetSessionIDFromContext(ctx); err == nil {
 		sess, e := s.SessManager.GetSession(sessionID)
 		if e != nil {
@@ -614,7 +614,7 @@ func (s *ImmuServer) getLoggedInUserdataFromCtx(ctx context.Context) (int64, *au
 	}
 
 	u, err := s.getLoggedInUserDataFromUsername(jsUser.Username)
-	return jsUser.DatabaseIndex, u, err
+	return int(jsUser.DatabaseIndex), u, err
 }
 
 func (s *ImmuServer) getLoggedInUserDataFromUsername(username string) (*auth.User, error) {
